New: Backend changes for new UI

+++ src/NzbDrone.Core/Jobs/ScheduledTask.cs
@@ -6,7 +6,7 @@ namespace NzbDrone.Core.Jobs
     public class ScheduledTask : ModelBase
     {
         public string TypeName { get; set; }
-        public double Interval { get; set; }
+        public int Interval { get; set; }
         public DateTime LastExecution { get; set; }
     }
 }
